requirements.txt
Django~=5.1.3
pip~=23.2.1
asgiref~=3.8.1
sqlparse~=0.5.1
requests~=2.32.3
certifi~=2024.8.30
urllib3~=2.2.3
idna~=3.10
djangorestframework~=3.15.2